Launch Screen View

+++ Let Me Do/Let Me Do/Views/ListView.swift
@@ -13,18 +13,24 @@ struct ListView: View {
     @State var isToggle : Bool = false
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){ item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.easeOut){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemViews()
+            } else{
+                List{
+                    ForEach(listViewModel.items){ item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.easeOut){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
         }
         .navigationTitle("Todo List 📝")
         .toolbar {
